+++ log/syslog/syslog_test.go
